config: Provision service so that user allowed modifications are preserved
User is able to change autoconnect value for a provisioned service.
This user modification must be preserved across ConnMan restarts.
This is done by loading the necessary minimal service settings
when provisioning a service via .config file. We also do not load
the service settings if we could provision the service so that
the options in .config file are not overwritten.