[connman] Enhance online check.
This enhances the online check by continusously checking for
connectivity if it fails. Previously it would check up to 12 times
and then give up. Don't give up, instead continue checking at the
maximum interval until it succeeds.

If the user changes the proxy configuration redo the online check.

Fixes issues with calling __connman_wispr_start() multiple times. If
there is a pre-existing wispr context it is cancelled, freed, and
recreated. However, no all network operations are cancelled or able to
be cancelled. Causes freed memory to be accessed when the network
operation completes.