service: Reset service to idle state when retrying after a failure
A failed service that is reconnected has to be set to idle state before
connecting it again. Changing state from failed to idle also has to be
allowed. Unless this is done, a service already in failed state won't
be able to detect that a subsequent reconnect also must re-trigger
a new reconnect attempt in __connman_service_ipconfig_indicate_state()
and service_indicate_state().

Fixes CM-622

Conflicts:
	connman/src/service.c