ipconfig: Do not disable kernel IPv6 support when disabling ipconfig
When ipconfig is disabled, it is not really necessary to disable
IPv6 support. Doing so will lead to issues when switching from
one service to another in the same network interface. User would
see that the IPv6 is disabled in kernel although user has IPv6
enabled in the service.

Thanks to Arjuna S <arjunasiva@gmail.com> for reporting this.