ipconfig: Define unique value for CONNMAN_IPCONFIG_TYPE_ALL
Make a distinction between all and unknown ipconfig type values. As a
result Update the code to handle CONNMAN_IPCONFIG_TYPE_ALL in switch
statements.

As the unknown and all ipconfig values are now semantically different,
modify __connman_service_is_connected_state() to return the combined
state of IPv4 and IPv6 when given a value of CONNMAN_IPCONFIG_TYPE_ALL.