[connman] ofono: Set CONNMAN_IPCONFIG_METHOD_AUTO when no address defined
It is possible that IPv6.Settings-properties does not include
ip-address, gateway or nameservers when the context is made active.
This is due underlying ofono-driver/ril sending the assigned ipv4-addresses
first and then the assigned ipv6-addresses (within 3 secs).

This happens especially only with dual/ipv4v6-protocol and Ofono does not
currently have method to signal the updated values to it's clients.

However there is a way to workaround this by setting the ipv6_method
as CONNMAN_IPCONFIG_METHOD_AUTO and letting the connman find about
those ipv6-addresses (ip-address, gateway, nameservers) by itself.