resolver: Remove resolver flush function as it is not needed
This commit cleans up the way the code sends DNS requests when
name servers are added or removed. Next commit will do the
actual flushing properly.

We need to check if fallback nameservers need to be added to
DNS server list when a DNS server is removed. This is now done
in remove_entries() function, earlier it was done in obsolete
connman_resolver_flush() function.