iptables: Keep the current counter values on changes
When new iptables are commited, the counters in the kernel are reset and
the values of the old tables are returned. By keeping track of the original
index of an iptable entry, the new table can be updated with the original
counter values.

When a policy of a builtin chain is changed, its counter value is reset.
When new rules are added, their counter is preset to 0.