build: Create directories before putting files in them
When automake's dependency tracking is enabled a mirror of the source
tree is created in the build directory as it writes the dependency
information.

However, if dependency tracking is disabled on an out-of-tree build
this mirror isn't created and it's possible that the .service files
can't be written as the directories they go into haven't been created
yet (racing against the compiles, which do create directories).