firewall: Add support for single rule add/remove operations
We like to add and remove rules while the firewall is up and running.
For example we need to insert per Session rule in the global NAT table.
That could also be implemented destroying the whole table and recreate
it when need but that is quite an overhead.

Instead of taking down the whole table down we add an API to add and
remove new rules during runtime.