session: Add SNAT rules per session
The policy routing doesn't work because:

(local process) -> (routing decision) -> [raw:output] ->
  contrack -> [mangle:output] -> (reroute check) ->
  [nat:output] -> [filter:output] -> ...

When the local process generates a packet, the network stack takes
the routing decision really early on. In our case earlier then we
are able to mark them with iptables. The marking happens in the
mangle table and therefore the policy routing table is not
considered at this point.

My naive assumption was that the reroute check would be done
unconditionally. Unfortunately, this is not true.

The reroute check is only done if the packet was modified in the
mangle output chain. So for example for the ICMP protocol:

net/ipv4/ping.c:
  ping_v4_sendmsg()
    ip_route_output_flow()

And the rerouting happens here:

net/ipv4/netfilter/iptable_mangle.c
  ipt_mangle_out()
    ip_route_me_harder()
      ip_route_output_key()
        ip_route_output_flow()

So let's reroute the packets via

iptables -t nat -A POSTROUTING -o eth0 -j SNAT --to-source 192.168.1.100

The policy routing table is not completely useless. For example the
process could mark its packet on its own. Another possible way is to
have a cgroup controller setting the SO_MARK or we teach the policy
routing table to match on sk_classid (cls cgroup controller).

Obviously, we don't need both the policy routing table and the NAT
rules. But I am reluctant to rip out the policy routing code at this
point. Maybe someone is using it already.

Reported by: Gianfranco Casanova <gianfranco.casanova@gmail.com>