service: Do not send PropertyChanged signals before service is announced
Send no service PropertyChanged signals before the service is announced
in ServicesChanged. Move the ServicesChanged adding right after
registering a service to D-Bus, as any ipconfig state updates in between
caused service PropertyChanged signals to be sent as the service was
missing from the new services hash for ServicesChanged signal.