iptables: Add basic --protocol parser support
The SNAT extension wants to know which protocol is used. We need
to pass in a valid ipt_entry with the protocol set for this reason.

The parser is keept minimal, which means it does not support all
fancy features as iptables offers. Though for our main purposes it
is enough which is

iptables-test -t nat -A POSTROUTING -o enp0s20u2u1 -j SNAT --to-source 192.168.101.6